Changed button clicks to change events via BoardGameHandler
